+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## 1.0.0
+## 1.1.0
+- feat: SvgWebImage, use WebView to convert svg to png, the display is the same as the browser.
 
-* SvgImage for Flutter Image.
-* CachedNetworkSvgSource for jovial_svg ScalableImageSource.
+## 1.0.0
+- SvgImage for Flutter Image.
+- CachedNetworkSvgSource for jovial_svg ScalableImageSource.